+++ Nakama.Tests/MatchTest.cs
@@ -129,7 +129,7 @@ namespace Nakama.Tests
 
             client1.Send(NMatchCreateMessage.Default(), (INMatch match) =>
             {
-                client2.Send(NMatchJoinMessage.Default(match.Id), (INMatchPresences matchPresences) =>
+                client2.Send(NMatchJoinMessage.Default(match.Id), (INMatch match2) =>
                 {
                     evt.Set();
                 }, (INError err) =>
@@ -174,7 +174,7 @@ namespace Nakama.Tests
                 joinedUserId = args.MatchPresence.Join[0].UserId;
                 evt2.Set();
             };
-            client2.Send(NMatchJoinMessage.Default(m.Id), (INMatchPresences matchPresences) =>
+            client2.Send(NMatchJoinMessage.Default(m.Id), (INMatch match) =>
             {
                 // No action.
             }, (INError err) =>
@@ -197,7 +197,7 @@ namespace Nakama.Tests
             client1.Send(NMatchCreateMessage.Default(), (INMatch match) =>
             {
                 m = match;
-                client2.Send(NMatchJoinMessage.Default(m.Id), (INMatchPresences presences) =>
+                client2.Send(NMatchJoinMessage.Default(m.Id), (INMatch match2) =>
                 {
                     evt1.Set();
                 }, (INError err) =>
@@ -244,7 +244,7 @@ namespace Nakama.Tests
             client1.Send(NMatchCreateMessage.Default(), (INMatch match) =>
             {
                 m = match;
-                client2.Send(NMatchJoinMessage.Default(match.Id), (INMatchPresences matchPresences) =>
+                client2.Send(NMatchJoinMessage.Default(match.Id), (INMatch match2) =>
                 {
                     evt1.Set();
                 }, (INError err) =>